chat/requirements.txt
pyaudio
requests
numpy
openwakeword==0.5.1
rpi_rf
python-vlc
pyradios
pyusb==1.0.2
click==7.0
RPi.GPIO
python-dotenv
pvporcupine
psutil
pydub